PCINTS-343: Speed up configuration reads.

+++ blotter-template/blade/DataSource/bin/start-jar.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 #
-# Start the Blotter Template Java Adapter
+# Start the Blotter Template Java Adapter.
 #
 # $1 - Path to java executable
 # $2 - Path to datasource config file
@@ -14,6 +14,14 @@ BLADENAME=@adapterName@
 
 if [ "$1" = "CONFREADER" ]; then
    shift
+
+   machine=`uname -m`
+   CONFREADER_BINARY="${CONFIG_BASE}/../tools/DataSource/bin/configreader-$machine"
+   if [[ -f $CONFREADER_BINARY ]]; then
+       "${CONFREADER_BINARY}" "$@" 2> /dev/null
+       exit $?
+   fi
+
    java -cp "${BINARY_ROOT}/lib/*" com.caplin.datasource.DataSource "$@"
    exit $?
 else
@@ -28,4 +36,3 @@ else
    fi
    echo $!
 fi
-
