+++ pricing-template/blade/DataSource/bin/start-jar.sh
@@ -3,8 +3,8 @@
 # Start the Pricing Template Java Adapter.
 #
 # $1 - Path to java executable
-# $2 - Path to datasource XML file
-# $3 - Path to fields XML file
+# $2 - Path to datasource config file
+# $3 - Path to fields config file
 # $4 - Java definitions ( optional )
 #
 # Returns the process id of the Java process.
@@ -14,6 +14,14 @@ BLADENAME=@adapterName@
 
 if [ "$1" = "CONFREADER" ]; then
    shift
+
+   machine=`uname -m`
+   CONFREADER_BINARY="${CONFIG_BASE}/../tools/DataSource/bin/configreader-$machine"
+   if [[ -f $CONFREADER_BINARY ]]; then
+       "${CONFREADER_BINARY}" "$@" 2> /dev/null
+       exit $?
+   fi
+
    java -cp "${BINARY_ROOT}/lib/*" com.caplin.datasource.DataSource "$@"
    exit $?
 else
